MxShop/requirements.txt
django

djangorestframework-jwt

djangorestframework
django-filter
django-crispy-forms
django-reversion
django-formtools
markdown
coreapi

django-cors-headers
django-redis

pillow
mysqlclient

future
httplib2
requests

six

xlwt
xlsxwriter

drf-extensions
pycryptodome
raven